Make it appear in launchers and app stores
Right now KMenuEdit has no AppStream metadata and has NoDisplay=true set
in its .desktop file. This means that most of the methods for a user to
find the app don't work. If the user doesn't have it installed but
searches for it using Discover, no results will be found.

However this app is a useful, user-friendly GUI app and there is no
real reason why it should be hidden in this way.

This commit makes it visible, adjusts its metadata a bit, and gives it
AppStream metadata so that it can be found by app stores like Discover.