follow HIG on UI label case
The HIG recommend sentence case for labels in front of controls, and
menu items generally fall under the rules that license title case.
kmenuedit generally follows these rules, but there are a small number if
cases where the wrong capitalization is used.

This change makes these labels consistent with the HIG.