Don't create new shortcut component when clearing shortcuts
When removing an application we want to remove any shortcuts associated
with them, so we set the shortcut to an empty QKeySequence.

When setting the shortcut and no such component exists we create
a kglobalaccel desktop file for it. That makes sense when we want to
*add* a shortcut, but not when removing them.

CCBUG: 448609
(cherry picked from commit 34ed6840b57dcd021970ff82b497cefac1c0888f)